updated the structural learning example

+++ src/structures.jl
@@ -3,7 +3,7 @@ struct Graph
     n::Int64
     m::Int64
     edgelist::Dict{Array{Int64, 1}, Real}
-    neighbors::Array{Array{Int32, 1}, 1}
+    neighbors::Array{Array{Int64, 1}, 1}
 end
 
 struct DirGraph
